Update camera100.py
not brilliant, but I can see your coming from an Arduino environment so you script quite different. Your setup didn't require outputting if the initial key word is set.  Do try to use keywords.  `global` keyword is used inside functions when you want to use the variable defined in the global scope.  Everything in the global scope is global you don't have to define it as global.  `from * import *` is cleaner and shorter in larger scripts.  I imagine you will utilize all the detections in each "frame", but for now you seem to be only looking for one detection in each frame.  Argurment Parser is OK, but if this is running on a device from boot, you would be better off with making a configuration file and reading from it at initialization, look into `.ini` files and `configparser`.  `sys.exit()` is pretty much the same as `exit()`.  Your `try/except` statement should really catch a specific `Exception`